--- a/maintainers/scripts/convert-to-import-cargo-lock/shell.nix
+++ b/maintainers/scripts/convert-to-import-cargo-lock/shell.nix
@@ -1,0 +1,5 @@
+with import ../../../. { };
+
+mkShell {
+  packages = [ rustc cargo clippy rustfmt ] ++ lib.optional stdenv.isDarwin libiconv;
+}