This PR proposes converting all Rust packages in Nixpkgs to using `importCargoLock` instead of `fetchCargoTarball` whenever possible, which [parses the lock file to generate independent FODs, per dependency, using the hashes from it](https://github.com/NixOS/nixpkgs/blob/master/pkgs/build-support/rust/import-cargo-lock.nix).

## Why now?

*Alternatively: "why isn't this a proper RFC?"*

On March 9th, 2023, [Rust 1.68.0 will be released](https://forge.rust-lang.org/#current-release-versions). With it, Cargo will introduce a change that will [break our FODs for packages that use Git dependencies](https://github.com/rust-lang/cargo/pull/11414).

We just **don't have time** to go through the RFC process, unless we want to fix all affected hashes.

@alyssais thought that this oppertunity was perfect to just vendor the lockfiles, and after some discussion, I agree.

## Benefits

1. We don't have to download the ~170 MiB registry each time we want to fetch a package's dependencies.
2. We can share crates between packages in the Nix store, as the resulting vendor directories just contain symlinks.
3. We can eventually do fancy things such as automatically include required libraries for crates into package's build inputs, or per-crate build artifact caching, which -- until now -- would require IFD.

## Drawbacks

1. Repo size inflation. The PoC conversion adds ~32.6 MiB to the repo's size. I believe this is justifiable because of the above space-saving changes.
2. Eval performance will naturally decrease. However, not by a lot:

|            stat            |     before     |     after      |       Δ       |  Δ%   |
|:---------------------------|---------------:|---------------:|:--------------|------:|
| **cpuTime**                |          73.86 |          74.78 | ↗ 0.92        | 1.24% |
| **envs-bytes**             |  2,999,282,224 |  3,040,232,472 | ↗ 40,950,248  | 1.37% |
| **envs-elements**          |    162,064,778 |    164,196,713 | ↗ 2,131,935   | 1.32% |
| **envs-number**            |    106,422,750 |    107,916,173 | ↗ 1,493,423   | 1.40% |
| **gc-heapSize**            |  9,681,682,432 |  9,883,009,024 | ↗ 201,326,592 | 2.08% |
| **gc-totalBytes**          | 17,666,947,056 | 17,862,112,352 | ↗ 195,165,296 | 1.10% |
| **list-bytes**             |    429,334,344 |    435,001,192 | ↗ 5,666,848   | 1.32% |
| **list-concats**           |      2,021,476 |      2,186,700 | ↗ 165,224     | 8.17% |
| **list-elements**          |     53,666,793 |     54,375,149 | ↗ 708,356     | 1.32% |
| **nrAvoided**              |    148,568,027 |    150,413,380 | ↗ 1,845,353   | 1.24% |
| **nrFunctionCalls**        |     88,721,432 |     90,120,887 | ↗ 1,399,455   | 1.58% |
| **nrLookups**              |     45,592,738 |     46,857,868 | ↗ 1,265,130   | 2.77% |
| **nrOpUpdateValuesCopied** |    248,764,456 |    250,934,019 | ↗ 2,169,563   | 0.87% |
| **nrOpUpdates**            |     10,805,375 |     10,967,351 | ↗ 161,976     | 1.50% |
| **nrPrimOpCalls**          |     70,154,155 |     70,923,773 | ↗ 769,618     | 1.10% |
| **nrThunks**               |    149,433,291 |    151,489,250 | ↗ 2,055,959   | 1.38% |
| **sets-bytes**             |  5,943,511,488 |  5,992,796,432 | ↗ 49,284,944  | 0.83% |
| **sets-elements**          |    341,296,918 |    344,142,502 | ↗ 2,845,584   | 0.83% |
| **sets-number**            |     30,172,550 |     30,407,275 | ↗ 234,725     | 0.78% |
| **sizes-Attr**             |             16 |             16 | 0             |       |
| **sizes-Bindings**         |             16 |             16 | 0             |       |
| **sizes-Env**              |             16 |             16 | 0             |       |
| **sizes-Value**            |             24 |             24 | 0             |       |
| **symbols-bytes**          |      1,828,318 |      1,842,339 | ↗ 14,021      | 0.77% |
| **symbols-number**         |        146,044 |        146,738 | ↗ 694         | 0.48% |
| **values-bytes**           |  4,585,180,872 |  4,639,916,904 | ↗ 54,736,032  | 1.19% |
| **values-number**          |    191,049,203 |    193,329,871 | ↗ 2,280,668   | 1.19% |

```bash
# pre-conversion:
________________________________________________________
Executed in   86.59 secs    fish           external
   usr time   73.89 secs    0.27 millis   73.89 secs
   sys time    7.61 secs    2.88 millis    7.61 secs

# post-conversion:
________________________________________________________
Executed in   87.57 secs    fish           external
   usr time   74.85 secs   55.14 millis   74.80 secs
   sys time    7.38 secs    7.21 millis    7.37 secs
```

## What about other languages?

*Alternatively: "you maintain a FOD-based builder, are you going to push vendoring their lockfiles?"*

Rust's lockfile format is uniquely suitable to this type of vendoring because they turn out to be pretty small (unlike npm or Yarn's), and we can use the hashes from them directly (unlike Go's).

All in all, this PR merely exists as a stopgap until [computed derivations](https://github.com/NixOS/rfcs/blob/master/rfcs/0092-plan-dynamism.md) are in a state where we can use them. Once they are, they will obviously be superior to anything vendoring can do.

Assuming the consensus is that we want to move forward with this, I will update this PR to include [my PoC branch](https://github.com/winterqt/nixpkgs/tree/import-cargo-lock-migration-poc), which is where you can take a look at what this will end up looking like. This branch converts 879 packages, which were all done with the script I've added in this PR. When the actual conversion happens, I'll go through and do the rest manually. I also need to update documentation.

I appreciate any and all thoughts on this, thanks!