@nixos-discourse 2023-02-19
This pull request has been mentioned on **NixOS Discourse**. There might be relevant details there:

https://discourse.nixos.org/t/migrate-all-rust-packages-to-importcargolock/25582/1