@figsoda 2023-02-19
I am not going to repeat the benefits of this change, overall this is a 👍 for me. I will just link a few related PRs

- #215629 is required for the treewide change, otherwise older lock files will fail to eval
- #215635 will ease the implementation of external tooling like [nix-update](https://github.com/Mic92/nix-update), which should (hopefully) make the transition rather painless. I am also going to work on `cargoLock` support for [nix-init](https://github.com/nix-community/nix-init).

> Cargo will introduce a change that will  break our FODs for packages that use Git dependencies.

Unrelated to this proposal, but we should probably backport this change to `importCargoLock`, as this is blocking updates of ruff and probably more packages in the future