@winterqt 2023-02-19
> Is this going to be a user-visible change in policy? If so, we should document it in the manual.

Yes, and I noted this in the proposal:

> I also need to update documentation.

(Admittedly, it was buried at the end, so I don't blame you for missing it.)