@Mic92 2023-02-19
Will just break our current checksums or will it render the old vendor mechanism useless?