@winterqt 2023-02-19
> cc @ryantm as this will also break automatic updates in ryantm-r...

Was gonna poke him if the overall opinion was largely positive.

> Will just break our current checksums or will it render the old vendor mechanism useless?

The Cargo change? It'll only invalidate some hashes, it won't break the mechanism entirely.