@alyssais 2023-02-19
> Speaking of, we should determine some sort of size threshold that warrants not vendoring the lock file... 1 MiB? 5 MiB?

Are there any packages we know of for which this might be an issue?  (Also compressed size is probably what matters — not sure how far we want to do down ease-of-check vs accuracy.)