@winterqt 2023-02-19
> > Cargo will introduce a change that will  break our FODs for packages that use Git dependencies.
> 
> Unrelated to this proposal, but we should probably backport this change to `importCargoLock`, as this is currently blocking updates of ruff and probably more packages in the future

[Even better, let's do it before them 😉](https://github.com/NixOS/nixpkgs/pull/217232)

> > Speaking of, we should determine some sort of size threshold that warrants not vendoring the lock file... 1 MiB? 5 MiB?
> 
> Are there any packages we know of for which this might be an issue?

I'm unsure off the top of my head, though I'm sure we can run the numbers on the PoC branch.

> (Also compressed size is probably what matters — not sure how far we want to do down ease-of-check vs accuracy.)

What do you mean by the latter, considering everything is losslessly compressed?