@alyssais 2023-02-20
> > (Also compressed size is probably what matters — not sure how far we want to do down ease-of-check vs accuracy.)
> 
> What do you mean by the latter, considering everything is losslessly compressed?

When considering whether a lockfile is too big, the most important thing to configure is how much size it permanently adds to the repo, which would be the size after being gzipped and maybe included in a packfile, rather than the size it temporarily adds to a checkout.  (Hopefully Cargo.lock files will compress extremely well when packed together, as there'll be a lot of duplication between them.)