@nixos-discourse 2023-03-02
This pull request has been mentioned on **NixOS Discourse**. There might be relevant details there:

https://discourse.nixos.org/t/2023-02-30-nixpkgs-architecture-team-meeting-30/25973/1