@tomberek 2023-03-05
I don't yet have an fully-formed opinion on this, but some thoughts

- Rust is the best candidate for something like this.
- We can take a smaller step by updating the FOD hashes and leaving the Cargo.lock vendoring to be done over time or at a later time.
- The size increase is significant, and while less for Rust that it might be for other languages, where do we draw the line? What is the policy?
- How does this affect the ability for people to learn what is best-practice for Rust packaging?
- What kind of mismatches can this cause?
- How does this affect contributions?
- Documentation of this being the Nixpkgs approach.