@figsoda 2023-03-05
> We can take a smaller step by updating the FOD hashes and leaving the Cargo.lock vendoring to be done over time or at a later time.

That is a possible solution. But the tree wide migration to `cargoLock` would not be much more immediate work, since that we already have the migration script and fixed a few shortcomings of `importCargoLock`. If we do plan on migrating everything to `importCargoLock` (which seems like the case given the positive responses so far), a tree wide migration would be easier in the long run IMO.

> The size increase is significant, and while less for Rust that it might be for other languages, where do we draw the line? What is the policy?

My understanding is that we are only doing this to Rust for now. It would be possible for node packages if we move something like npmlock2nix into nixpkgs, but the `package.json` files are a little large. Though I still think it would be a better approach than node2nix and yarn2nix, whether it would replace `buildNpmPackage` and `nodePackages` is up for future discussion. I don't think any other build system has support similar to `importCargoLock`, please correct me if I missed anything.

> How does this affect the ability for people to learn what is best-practice for Rust packaging?
>
> How does this affect contributions?

I think we should enforce new packages to use `cargoLock` when possible, and have this be the recommended approach to package Rust packages.

> What kind of mismatches can this cause?

I'm assuming you mean regression? It should be fairly minimal, as `importCargoLock` is already widely used outside nixpkgs, and we have fixed/are fixing [a](https://github.com/NixOS/nixpkgs/pull/218472) [few](https://github.com/NixOS/nixpkgs/pull/215629) [issues](https://github.com/NixOS/nixpkgs/pull/217232) with it. This PR is targeting master right now, but it should probably go to staging instead, and we can catch any regressions that happen before it gets into master.