@nixos-discourse 2023-03-06
This pull request has been mentioned on **NixOS Discourse**. There might be relevant details there:

https://discourse.nixos.org/t/2023-03-06-nixpkgs-architecture-team-meeting-31/26070/1