@infinisil 2023-03-06
We discussed this in [todays Nixpkgs Architecture Team](https://discourse.nixos.org/t/2023-03-06-nixpkgs-architecture-team-meeting-31/26070#vendored-rust-cargolocks-5) meeting. We think that the size increase is too much and recommend instead just updating the `cargoSha256`'s.